Renamed files/folders
controllers -> routes
services -> models
